fix a memory leak in the list tcs function for a test program
When compiling under address sanitizers, the test program would
show a leak memory result from trying to list the test cases.
This is caused because the atf_tp_get_tcs returns an allocated
array that is not subsequently freed.

Unfortunately, the secondary issue is that test program output
can not be parsed by kyua because of the ASAN output at the end
of the "-l" invocation. So, the testcases get listed like:

t_foo:__test_cases_list__